windmill-analysis/requirements.txt
flask
opencv-python
numpy
scipy
matlabengine
opencv-contrib-python
Pillow
matplotlib
pyinstaller